pyproject.toml
[project]
name = "daneel"
version = "0.1.0"
description = "Python helper functions for agentic coding assistants"
authors = [
    {name = "Daneel Project"}
]
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0",
    "pexpect>=4.8",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "mypy>=1.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project.scripts]
daneel = "daneel:main"

[tool.ruff]
line-length = 88
target-version = "py311"

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]